move View into model
